Node v16 Update!
Notable Changes
- Updated to node-redis 3.1.2
- Updated to redlock 4.2.0
- Updated to okanjo-app 3.x

- Added GH actions for CI
- Removed .eslintrc.json .eslintignore .travis.yml
- RedisService.js: Updated for okanjo-app v3
- package.json:
  - Bumped to v4.0.0
  - Updated all dev dependencies to latest
  - Moved mocha, eslint configs

The node-redis package was not updated to latest due to the complete mess they made out of the module in v4.
- No longer supports Redis v4 (breaking)
- Functions now return promises (yay!)
- Configuration schema has totally changed (breaking)
- Callbacks can be re-enabled with legacyMode (cool)
- Renamed functions to pascalCase or UPPERCASE (breaking, also very not cool)

In order to upgrade to node-redis v4, the entirety of the module would need to be refactored, as would any module using redis. It would seem that another major release in the future could address this, but would come at the cost of a major redis overhaul in all affected apps.